backend/requirements.txt
Flask>=2.0
python-dotenv>=0.21
flask-cors>=3.0
gunicorn>=20.1
requests>=2.31.0
